Separate home, quote and display snapshot test files
