+++ src/_tests_/app.test.js
@@ -4,8 +4,6 @@ import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 import Button from '../components/Button';
 import ButtonPanel from '../components/ButtonPanel';
 import Navbar from '../components/Navbar';
-import operate from '../logic/operate';
-import calculate from '../logic/calculate';
 
 Enzyme.configure({ adapter: new Adapter() });
 
@@ -20,44 +18,4 @@ describe('<App />', () => {
     expect(wrapper).toBe(19);
   });
 
-  it('returns the addition of two nums', () => {
-    const result = operate(4, 5, '+');
-    expect(result).toEqual('9');
-  });
-
-  it('calculates the product of two digits', () => {
-    const result = calculate({
-      total: '6', next: '5', operation: 'x',
-    }, '=');
-    expect(result).toEqual({
-      total: '30', next: null, operation: null,
-    });
-  });
-
-  it('does not calculate the wrong product of two digits', () => {
-    const result = calculate({
-      total: '10', next: '5', operation: 'x',
-    }, '=');
-    expect(result).not.toEqual({
-      total: '30', next: null, operation: null,
-    });
-  });
-
-  it('does not calculate the wrong difference of two digits', () => {
-    const result = calculate({
-      total: '100', next: '50', operation: '-',
-    }, '=');
-    expect(result).not.toEqual({
-      total: '40', next: null, operation: null,
-    });
-  });
-
-  it('calculates the dividend of two digits', () => {
-    const result = calculate({
-      total: '10', next: '5', operation: '÷',
-    }, '=');
-    expect(result).toEqual({
-      total: '2', next: null, operation: null,
-    });
-  });
 });
