+++ src/_tests_/display.test.js
@@ -2,8 +2,6 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import Display from '../components/Display';
 import Button from '../components/Button';
-import Quote from '../components/Quote';
-import Home from '../components/Home';
 
 it('renders button correctly', () => {
   const tree = renderer.create(<Button name="AC" clickHandler={jest.fn()} />);
@@ -12,15 +10,5 @@ it('renders button correctly', () => {
 
 it('renders display correctly', () => {
   const tree = renderer.create(<Display result="0" next="" />);
-  expect(tree).toMatchSnapshot();
-});
-
-it('renders quotes page', () => {
-  const tree = renderer.create(<Quote />);
-  expect(tree).toMatchSnapshot();
-});
-
-it('renders home page', () => {
-  const tree = renderer.create(<Home />);
   expect(tree).toMatchSnapshot();
 });
