+++ src/_tests_/home.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import Home from '../components/Home';
+
+it('renders home page', () => {
+  const tree = renderer.create(<Home />);
+  expect(tree).toMatchSnapshot();
+});
