+++ src/_tests_/quote.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import Quote from '../components/Quote';
+
+it('renders quotes page', () => {
+  const tree = renderer.create(<Quote />);
+  expect(tree).toMatchSnapshot();
+});
